pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "semgrepai"
version = "0.2.0"
description = "AI-powered Semgrep vulnerability validator with async processing and web UI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "hardw00t"}
]
keywords = ["security", "semgrep", "vulnerability", "ai", "llm", "scanner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    # Core scanning
    "semgrep>=1.100.0",

    # LLM providers
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.4.0",
    "langchain-ollama>=0.3.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",

    # Vector database
    "chromadb>=0.5.0",

    # Async utilities
    "aiolimiter>=1.1.0",
    "tenacity>=8.2.0",

    # Web API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",

    # CLI and output
    "typer>=0.12.0",
    "rich>=13.5.0",
    "jinja2>=3.1.0",

    # Configuration and validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # System utilities
    "psutil>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",  # For testing FastAPI
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

ollama = [
    "langchain-ollama>=0.3.0",
]

openrouter = [
    "langchain-openai>=1.0.0",  # OpenRouter uses OpenAI API format
]

[project.scripts]
semgrepai = "semgrepai.cli:app"

[project.urls]
Homepage = "https://github.com/hardw00t/simple_semgrepAI"
Repository = "https://github.com/hardw00t/simple_semgrepAI"
Issues = "https://github.com/hardw00t/simple_semgrepAI/issues"

# Hatch build configuration
[tool.hatch.build.targets.wheel]
packages = ["semgrepai"]

[tool.hatch.build.targets.wheel.force-include]
"semgrepai/templates" = "semgrepai/templates"
"semgrepai/web/frontend/dist" = "semgrepai/web/frontend/dist"

[tool.hatch.build.targets.sdist]
include = [
    "semgrepai/**/*",
    "tests/**/*",
    "Makefile",
    "README.md",
    "LICENSE",
]

# UV-specific configuration
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "Q"]
ignore = ["E501"]  # Line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
